Add files via upload
Did Data Preprocessing, analyzed the columns, split the dataset into train and test with stratified distribution and readied the data to be transformed, after which, it can be used to train Machine Learning Algorithms. Since our target variable is numerical, and our task is to predict the power consumption based on the given variables, we can use regression algorithms to build the model.